requirements.txt
deepgram-sdk>=3.0.0
pyaudio>=0.2.14
pynput>=1.7.6
python-dotenv>=1.0.0
